requirements.txt
fastapi
uvicorn
httpx
python-dotenv
sqlalchemy
fpdf
google-generativeai
